completed error handling for DELETE method

+++ models/app.models.js
@@ -188,5 +188,26 @@ exports.postById = (article_id, username, body) => {
 }
 
 exports.deleteCommentById = (comment_id) => {
-    return db.query(`DELETE FROM comments WHERE comment_id = $1`, [comment_id])
+    const regex = /\d/g
+    if(regex.test(comment_id) === false) {
+        return Promise.reject({
+            status: 400, message: 'Please enter a valid comment_id' 
+        })
+    }
+
+    return db.query(`SELECT * FROM comments WHERE comment_id = ${comment_id}`)
+    .then((comments) => {
+        if (comments.rows.length === 0) {
+            return Promise.reject({
+                status: 404, message: `No comment found for comment_id: ${comment_id}`
+            })
+        }
+    })
+    .then(() => {
+        return db.query(`DELETE FROM comments WHERE comment_id = $1`, [comment_id])
+    })
+        
+    
+
+    
 }
